# docs/swagger.yaml
basePath: /swagger/aibox-service
definitions:
  common.Page:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  model.Aibox_device:
    properties:
      build_time_str:
        description: 设备构建时间
        type: string
      created_by:
        description: created_by
        type: string
      created_time:
        description: created_time
        type: string
      id:
        description: 设备ID
        type: string
      ip:
        description: 设备IP地址
        type: string
      latest_heart_beat_time:
        description: 最近心跳时间
        type: string
      name:
        description: 设备名称
        type: string
      status:
        description: 设备状态(0:离线，1:在线)
        type: integer
      updated_by:
        description: updated_by
        type: string
      updated_time:
        description: updated_time
        type: string
    type: object
...
info:
  contact: {}
  description: aibox-service API
  title: aibox-service API
  version: "1.0"
paths:
...
  /aibox-device:
    get:
      description: query objects
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: ip
        in: query
        name: ip
        type: string
      - description: build_time_str
        in: query
        name: build_time_str
        type: string
      - description: latest_heart_beat_time
        in: query
        name: latest_heart_beat_time
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Aibox_device'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: query objects
      tags:
      - AI盒子设备
    post:
      consumes:
      - application/json
      description: save
      parameters:
      - description: object
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Aibox_device'
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Aibox_device'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: save
      tags:
      - AI盒子设备
...
  /aibox-device/{id}:
    delete:
      description: delete
      parameters:
      - description: 实例id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Aibox_device'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: delete
      tags:
      - AI盒子设备
  /aibox-device/batch-delete:
    post:
      consumes:
      - application/json
      description: batch delete
      parameters:
      - description: id array
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch delete
      tags:
      - AI盒子设备
  /aibox-device/batch-upsert:
    post:
      consumes:
      - application/json
      description: batch update
      parameters:
      - description: objects array
        in: body
        name: entities
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch update
      tags:
      - AI盒子设备
  /aibox-device/page:
    get:
      description: page query, _page(from 1 begin), _page_size, _order, and others
        fields, status=1, name=$like.%CAM%
      parameters:
      - description: current page
        in: query
        name: _page
        required: true
        type: integer
      - description: page size
        in: query
        name: _page_size
        required: true
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: ip
        in: query
        name: ip
        type: string
      - description: build_time_str
        in: query
        name: build_time_str
        type: string
      - description: latest_heart_beat_time
        in: query
        name: latest_heart_beat_time
        type: string
      - description: status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/model.Aibox_device'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: page query
      tags:
      - AI盒子设备
...
swagger: "2.0"